requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-core==4.3.0
MarkupSafe==1.0
nbformat==4.4.0
numpy==1.13.2
pandas==0.20.3
plotly==2.0.15
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
scipy==1.0.0rc1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
Werkzeug==0.12.2
WTForms==2.1
